refactor(get-public-link-url.ts): add public link API
Added getPublicLinkUrl function to handle public link creation via
GraphQL mutation. Updated related imports and usages.
