+++ src/views/app/detail-panel/edit/parts/convert-to-smartlink-modal.tsx
@@ -9,7 +9,7 @@ import { Container, Text } from '@zextras/carbonio-design-system';
 import { ModalFooter, ModalHeader } from '@zextras/carbonio-ui-commons';
 import { useTranslation } from 'react-i18next';
 
-import { getPublicUrl } from './smart-link-utils';
+import { getPublicLinkUrl } from 'api/get-public-link-url';
 import { uploadToFiles } from 'api/upload-file-to-files';
 import { useEditorText } from 'store/editor/hooks';
 import { addSmartLinksToText } from 'ui-actions/utils';
@@ -28,9 +28,9 @@ export const ConvertToSmartlinkModal = ({
 	const { getText, setText } = useEditorText(editorId);
 	const onConfirm = useCallback(async () => {
 		const uploadToFilesResponse = await uploadToFiles(files[0]);
-		const publicUrlResponse = await getPublicUrl(uploadToFilesResponse.data.nodeId);
+		const publicLinkUrl = await getPublicLinkUrl(uploadToFilesResponse.data.nodeId);
 		const textWithLinks = addSmartLinksToText({
-			publicLinkUrl: publicUrlResponse.data.data.createLink.url,
+			publicLinkUrl: publicLinkUrl.data.data.createLink.url,
 			text: getText(),
 			filename: files[0].name
 		});
